pyproject.toml
[tool.poetry]
name = "sil"
version = "0.1.0"
description = "This repository include the code for the paper \"Enriching Social Science Research via Survey Item Linking\"."
authors = ["e-tornike <20404466+e-tornike@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.9,<3.11"
datasets = "^2.18.0"
evaluate = "^0.4.1"
numpy = "^1.26.4"
torch = "<2.0.0"
accelerate = "^0.27.2"
scikit-learn = "^1.4.1.post1"
sentence-transformers = "^2.5.1"
lightning = "^2.2.1"
faiss-gpu = "^1.7.2"
typer = "^0.9.0"
torchmetrics = "^1.3.1"
optuna = "^3.5.0"
jupyter = "^1.0.0"
jsonlines = "^4.0.0"
lingua-language-detector = "^2.0.2"
skops = "^0.9.0"
huggingface-hub = "^0.22.2"
tokenizers = "^0.15.2"
transformers = "4.37.2"
ranx = "^0.3.19"
tinycss2 = "1.2.1"
qdrant-haystack = "<2.0.0"
deepsig = "^1.2.8"
farm-haystack = {git = "https://github.com/deepset-ai/haystack.git", rev = "v1.25.x", develop = true}
quaterion-models = "^0.1.19"
quaterion = "^0.1.35"
opensearch-py = ">=2"
nltk = "^3.8.1"
langdetect = "^1.0.9"
nlpaug = "^1.1.11"
scienceplots = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.357"
pynvim = "^0.5.0"
ruff = "^0.3.5"
isort = "^5.13.2"


[tool.poetry.group.exp_icl.dependencies]
outlines = "^0.0.34"


[tool.poetry.group.data_s44k.dependencies]
doc2json = {git = "https://github.com/allenai/s2orc-doc2json.git"}



[tool.poetry.group.data_gsim.dependencies]
openai = "<1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"